Major changes in the application including DAO pattern introduction, data source initialization, Threading model changes.
Removed IFacilityProvider interface, created DAO pattern for the Facilities instead.
Added DAO implementation for the SQLite database, isolated storage model from the data itself.
Added FacilityDAOHelper and its SQLite implementation for the data initialization purpose, which is currently done in the separate AsyncTask on the start of the application, instead of doing it in the class derived from the SQLiteOpenHelper. User is properly informed about the application state.
Changed querying thread model. Instead of spawning new Thread each time the query is done, there is single dedicated thread with Handler exposed for that purpose. Improvements in requerying included.